Refactor slamdunk-all-paired process
